+++ bytes/doc.go
@@ -1,0 +1,2 @@
+// Package bytes provides additional tools for working with byte slices.
+package bytes