+++ Changes.md
@@ -1,3 +1,9 @@
+## WIP  TBD
+
+ * Adding `bytes.ContainsOnly`, `bytes.FromRange`, `bytes.Reverse`, `bytes.Indent`.
+ * Adding `strings.Indent`.
+ * Corrected documentation for `deferred.Error` which gave incorrect guidance on how to use it.
+
 ## v0.7.0  2024-06-13
 
  * :boom: Breaking Change :boom:: Now requires Go 1.20.